[hdf5] Split archive.cpp into 5 parts
archive_read.cpp is by far the slowest part to compile.
